Extract bulk inventory mutation engine
