+++ src/mtg_source_stack/inventory/operations/__init__.py
@@ -1,0 +1,1 @@
+"""Focused inventory write-operation modules."""
